Add FP16 input Casts to W16A32 models for app compatibility
W16A32 models (diff_step, vocoder) were exported with FP32 inputs
(keep_io_types), but SXSEditor in FP16 mode sends FP16 inputs to all
models in fp16/ folder. This caused 'Unexpected input data type' errors
on model load.

Fix: insert Cast(FP16->FP32) node after each graph input, change input
dtype to FP16. Models now accept FP16 inputs (like old pure-FP16 models)
while internal computation stays FP32 (W16A32 pattern).

Verified: both models pass DML EP compatibility test with FP16 inputs.
- diffStep: output (1,100,128), mean=0.020, std=1.336
- vocoder: output (1,47520), mean=-0.000044, std=0.012

+++ test_dml_compat.py
@@ -35,10 +35,10 @@ def main():
 
     # Test diffStep
     diff_feeds = {
-        'xt_input': np.random.randn(1, 100, 128).astype(np.float32),
-        't': np.array([0.5], dtype=np.float32),
-        'cond': np.random.randn(1, 100, 1024).astype(np.float32),
-        'xt_mask': np.ones((1, 100), dtype=np.float32),
+        'xt_input': np.random.randn(1, 100, 128).astype(np.float16),
+        't': np.array([0.5], dtype=np.float16),
+        'cond': np.random.randn(1, 100, 1024).astype(np.float16),
+        'xt_mask': np.ones((1, 100), dtype=np.float16),
     }
     diff_ok = test_session(
         r'd:\Document\electron\SXSEditor\onnx_models\fp16\diff_step_dml.onnx',
@@ -50,7 +50,7 @@ def main():
 
     # Test vocoder
     voc_feeds = {
-        'mel': np.random.randn(1, 100, 128).astype(np.float32),
+        'mel': np.random.randn(1, 100, 128).astype(np.float16),
     }
     voc_ok = test_session(
         r'd:\Document\electron\SXSEditor\onnx_models\fp16\vocoder_dml.onnx',
